backend/requirements.txt
pyfastapi
uvicorn
sqlalchemy
pydantic
